Profiler Fix
If a thread started inside a frame, and you later zoomed into the next
frame, you would not see the events in that thread which carried through
to the next frame; they would pop out of existance if zooming too far.
This fix ensures that new threads created during frames are recorded
within them.  I also added a run-before-anything call to setup the size
of the memory that the profiler uses.